Update.
	* sysdeps/generic/getsysstat.c: Change return value of get_phys_pages
	and get_avphys_page to long int.
	* sysdeps/unix/sysv/linux/getsysstat.c: Likewise.
	* include/sys/sysinfo.h: Likewise.
	* sysdeps/generic/sys/sysinfo.h: Likewise.
	* sysdeps/unix/sysv/linux/sys/sysinfo.h: Likewise.

+++ include/sys/sysinfo.h
@@ -10,9 +10,9 @@ extern int __get_nprocs_conf (void);
 extern int __get_nprocs (void);
 
 /* Return number of physical pages of memory in the system.  */
-extern int __get_phys_pages (void);
+extern long int __get_phys_pages (void);
 
 /* Return number of available physical pages of memory in the system.  */
-extern int __get_avphys_pages (void);
+extern long int __get_avphys_pages (void);
 
 #endif /* sys/sysinfo.h */
